pyscan split - check_dir
close #113
