+++ movie_plist/data/pyscan.py
@@ -1,11 +1,12 @@
 import os
 import re
-import sys
 import time
 from pathlib import Path
 from typing import Generator, Tuple
 
-from movie_plist.conf.global_conf import CFG_FILE, MOVIE_SEEN, MOVIE_UNSEEN
+from movie_plist.conf.global_conf import MOVIE_SEEN, MOVIE_UNSEEN
+
+from .check_dir import get_dir_path
 
 
 def create_dicts() -> None:
@@ -61,7 +62,6 @@ def _unknow_dirs() -> Generator[Path, None, None]:
     ((f.stat().st_mtime, f) for f in directory.iterdir())
     """
     _scan_dir = get_dir_path()
-    scan_dir_has_movies(_scan_dir)
 
     _json_movies = {**MOVIE_SEEN, **MOVIE_UNSEEN}
 
@@ -84,69 +84,10 @@ def _open_right_file(file_with_url: Path) -> str:
     if line_with_url:
         return line_with_url.group(0)[4:]
 
-    raise Exception(f'Please check {file_with_url} file. Path and content')
+    desktop_file = str(file_with_url.name)
+
+    raise Exception(f'Please check {desktop_file} file. Path and content')
 
 
 def mk_title_year(root_path: str) -> str:
     return root_path.rpartition('/')[-1]
-
-
-class InvalidPath(Exception):
-    pass
-
-
-def read_path() -> str:
-    content = Path(CFG_FILE)
-    cfg_file_path = content.read_text()
-
-    if not os.path.isdir(cfg_file_path):
-        raise InvalidPath('Invalid path in movie_plist.cfg file.')
-
-    return cfg_file_path
-
-
-def write_path(cfg_file_path: str) -> str:
-    if not os.path.isdir(cfg_file_path):
-        raise InvalidPath('Invalid path. Please try again.')
-
-    w_path = Path(CFG_FILE)
-    w_path.write_text(cfg_file_path)
-
-    return cfg_file_path
-
-
-def get_dir_path() -> str:
-    if os.path.isfile(CFG_FILE):
-        path_dir_scan = read_path()
-    else:
-        get_dir_scan = input(" Do the scan in which directory ? ")
-        path_dir_scan = write_path(get_dir_scan)
-
-    return path_dir_scan
-
-
-def scan_dir_has_movies(scan_dir: str):
-    # tem que fazer uma checagem melhor
-    for _, _, filename in os.walk(scan_dir):
-        for file in filename:
-            if file.endswith('.desktop'):
-                return True
-
-    from PyQt5.QtWidgets import QMessageBox, QApplication  # pylint: disable-msg=E0611
-
-    app = QApplication(['0'])  # noqa: F841
-
-    msg = QMessageBox()
-    msg.setIcon(QMessageBox.Warning)
-    msg.setWindowTitle("Empty Directory")
-
-    text = """
-        The directory scanned seems empty.
-        Please check the directory
-         """ + scan_dir
-
-    msg.setText(text)
-    msg.setStandardButtons(QMessageBox.Ok)
-    msg.exec_()
-
-    sys.exit('1')
